Fix issue with scaling using responsive: true
Previously the gallery would maintain whatever aspect ratio it happened
to load with. This would cause a problem for galleries that occupy the
full browser width. Loading the gallery in a large browser window and
then scaling the window down resulted in a smaller gallery size
(occupying less of the available window) than what the user saw on a
reload at that smaller window size.

This patch fixes the issue by maintaining a separate variable to store
the user-provided aspect ratio (which should still be respected), as
opposed to the calculated ratio based on the stage dimensions at load
time.