This line that modifies the transform property for individual images seems to make the images appear above the image layer on wekbit browsers with touch under specific circumstances. For example, take the Galleria Classic theme example for Galleria 1.3.5. Give galleria-layer a background-color. Open Chrome and emulate an iPad 3/4. Open up / refresh the theme example. Slowly drag the image to go to the next one.
Expected: You should see the background-color of the galleria-layer on top of the next image as you drag across.
Actual Result: The background layer seems to go behind the image.

This also seems to happen if you have a small number of images (like 3) and it wraps around from the end to the beginning or vice-versa by clicking the image-nav-left and image-nav-right divs.