Fix mobile device manipulation
Fixed issue https://github.com/aino/galleria/issues/315

I copied over and modified some code from the internal _run() function to the push() function which fixed this issue for me. Basically what this does is add the pushed image(s) to the _controls.slides array if swipe is enabled. This of course assumes you are using the galleria.push() method to push additional images to the Slideshow after the Slideshow is ready.